pyproject.toml
[project]
name = "building-ai-agents-with-langgraph"
version = "1.0.0"
description = "A multi-agent system built with LangGraph, LangChain, and OpenAI."
authors = [{name = "Sajal Sharma", email = "contact@sajalsharma.com"}]
license = "MIT"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jupyter>=1.1.1",
    "openai>=1.51.2",
    "python-dotenv>=1.0.1",
    "jupyterlab>=4.2.5",
    "langchain>=0.3.3",
    "langchain-openai>=0.2.2",
    "langchain-community>=0.3.2",
    "langgraph>=0.2.35",
    "langchain-experimental>=0.3.2",
    "matplotlib>=3.9.2",
    "tavily-python>=0.3.3",
    "alpha_vantage>=2.3.1",
    "langchain-tavily>=0.2.11",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["agents*", "graph*", "tools*"]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
